+++ src/main/kotlin/com/autonomousapps/internal/utils/collections.kt
@@ -1,5 +1,7 @@
 package com.autonomousapps.internal.utils
 
+import org.w3c.dom.Node
+import org.w3c.dom.NodeList
 import java.util.*
 import kotlin.collections.HashSet
 
@@ -32,3 +34,19 @@ internal inline fun <T, R> Iterable<T>.flatMapToSet(transform: (T) -> Iterable<R
 }
 
 internal fun <T> Iterable<T>.collectionSizeOrDefault(default: Int): Int = if (this is Collection<*>) this.size else default
+
+internal inline fun <R> NodeList.mapNotNull(transform: (Node) -> R?): List<R> {
+  val destination = ArrayList<R>(length)
+  for (i in 0 until length) {
+    transform(item(i))?.let { destination.add(it) }
+  }
+  return destination
+}
+
+internal inline fun <R> NodeList.map(transform: (Node) -> R): List<R> {
+  val destination = ArrayList<R>(length)
+  for (i in 0 until length) {
+    destination.add(transform(item(i)))
+  }
+  return destination
+}